requirements.txt
torch
torchtext
transformers
datasets
rich
wandb
hydra-core
lightning
evaluate
sentencepiece
seqeval